langfuse/requirements.txt
langfuse>=2.0
openai>=1.0
python-dotenv>=1.1.0
httpx>=0.28.0
pydantic>=2.10.0
rich>=13.0
